backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Momo website backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "modal>=0.72.10",
    "openai>=1.59.7",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
